Improved depot history but still not working

+++ server/api/depot_balance_apis.py
@@ -15,7 +15,7 @@ class User(BaseModel):
     disabled: bool | None = None
 
 class CashTransfer(BaseModel):
-    balance_sum: Decimal
+    balance_sum: float
     transaction_type: str
